modify & correct

+++ minuku2-extended/src/main/java/edu/umich/si/inteco/minuku/streamgenerator/FreeResponseQuestionStreamGenerator.java
@@ -76,7 +76,7 @@ public class FreeResponseQuestionStreamGenerator extends AndroidStreamGenerator<
         return true;    }
 
     @Override
-    public long getUpdateFrequency() {
+    public double getUpdateFrequency() {
         return -1;
     }
