+++ minuku2-extended/src/main/java/edu/umich/si/inteco/minuku/streamgenerator/SemanticLocationStreamGenerator.java
@@ -97,7 +97,7 @@ public class SemanticLocationStreamGenerator
     }
 
     @Override
-    public long getUpdateFrequency() {
+    public double getUpdateFrequency() {
         return -1;
     }
